% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reciprocal.R
\name{reciprocal}
\alias{reciprocal}
\title{Reciprocals}
\usage{
reciprocal(x, a)
}
\arguments{
\item{x}{The vector to be transformed (squared/cubed/etc.)}

\item{a}{The power to be calculated to, for the transformation}
}
\value{
A vector \code{x} that is the inverse of the power of \code{a}.
}
\description{
This function calculates the inverse power of a vector
}
\details{
This function calculates the inverse or "reciprocal" of powers, that is 1/X^a

For example, the inverse of 10 to the power of 2 is 1/100 = 0.01

This function also returns an informative error when a non-numeric input is given.
}
\examples{
reciprocal(-2, 3)
reciprocal(1:10, 2)
}
